requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Flask-Login==0.5.0
Flask-WTF==0.15.1
gunicorn==20.1.0
requests==2.26.0
python-dotenv==0.19.0
email-validator==1.1.3
Werkzeug==2.0.1
PyMySQL==1.0.2 